fix: 2차 세팅 변경
