--- idearly-fe/.prettierrc
@@ -1,10 +1,0 @@
-{
-  "singleQuote": true,
-  "parser": "typescript",
-  "semi": false,
-  "useTabs": false,
-  "tabWidth": 2,
-  "printWidth": 120,
-  "arrowParens": "always",
-  "organizeImports": "true"
-}
