--- idearly-fe/src/layout/Layout.tsx
@@ -1,13 +1,0 @@
-import { Outlet } from 'react-router-dom'
-import { Header } from './Header/Header'
-
-const Layout = () => {
-  return (
-    <>
-      <Header />
-      <Outlet />
-    </>
-  )
-}
-
-export default Layout
