--- idearly-fe/src/styles/theme.ts
@@ -1,20 +1,0 @@
-import { DefaultTheme } from '@emotion/react'
-
-const theme: DefaultTheme = {
-  fontSizes: {
-    xxs: '12px',
-    xs: '13px',
-    sm: '14px',
-    base: '16px',
-    md: '18px',
-    lg: '24px',
-  },
-  colors: {
-    black: '#000',
-    dark: '#191a20',
-    primary: '#3f4150',
-    secondary: '#8c8d96',
-  },
-}
-
-export default theme
