+++ idearly/src/App.tsx
@@ -1,0 +1,14 @@
+import { RouterProvider } from "react-router-dom";
+import router from "./routes/routing";
+import GlobalStyle from "./styles/global";
+
+function App() {
+  return (
+    <div>
+      <GlobalStyle />
+      <RouterProvider router={router} />
+    </div>
+  );
+}
+
+export default App;
