+++ idearly/src/layout/Layout.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from "react-router-dom";
+import { Header } from "./Header/Header";
+
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
