+++ idearly/src/pages/AlgorithmSolvingPage/AlgorithmSolvingPage.tsx
