+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmNav/AlgorithmNav.tsx
