+++ idearly/src/pages/AlgorithmSolvingPage/components/index.ts
