+++ idearly/src/pages/AlgorithmSolvingPage/layout/AlgorithmLayout.tsx
