+++ idearly/src/pages/DetailPage/DetailPage.tsx
