+++ idearly/src/pages/WaitingPage/WaitingPage.tsx
